pyproject.toml
[tool.poetry]
name = "acs-crm"
version = "0.1.0"
description = "Managing client info and tasks"
authors = ["Konstantinos Vitiniotis <kostasvit@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}
django-extensions = "^3.2.3"
django-import-export = "^4.3.1"
django-filter = "^24.3"
django-widget-tweaks = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
